# EventHub-BE/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "EventHub BE",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5678",
    "basePath": "/",
    "paths": {
...
        "/auth/regis": {
            "post": {
                "description": "Register a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Regis"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "Register payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Regis"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "registration successful",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "400": {
                        "description": "password length error / bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "409": {
                        "description": "user or email already exists",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "binding error / internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
        "/events/upcoming": {
            "get": {
                "description": "Get list of all upcoming events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events upcomming"
                ],
                "summary": "Get upcoming events",
                "responses": {
                    "200": {
                        "description": "success get upcoming events",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.UpcomingEvent"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "failed to get upcoming events",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
        "/testimonials": {
            "get": {
                "description": "Get list of all testimonials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get all testimonials"
                ],
                "summary": "Get all testimonials",
                "responses": {
                    "200": {
                        "description": "testimonials retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Testimonial"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Create a new testimonial (requires authentication)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Create testimonials"
                ],
                "summary": "Create a new testimonial",
                "parameters": [
                    {
                        "description": "Create testimonial payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTestimonial"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "testimonial created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "failed to create testimonial",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateTestimonial": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "dto.Res": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "token": {}
            }
        },
...
        "model.Regis": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.Testimonial": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "job": {
                    "type": "string"
                },
                "profile": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.UpcomingEvent": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "attendees": {
                    "type": "integer"
                },
                "capacity": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "event_format": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "images": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}